pyproject.toml
[project]
name = "llm-fight-club"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.80.11",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]
